Add files via upload
Please note there were 2 implementations of the stack, 2 implementations of the expression function (I tried to implement an expression function that takes parentheses in consideration for the proper order of operation (not completed though)) I also made requirements 5 and 6 both in the same project of Data_structures_linked_lists with requirement 6 in main while 5 in the files. I did not check all linked list function but they seem to work fine I will recheck though. Please wait for another update as well